requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.3
cssselect2==0.7.0
daphne==4.0.0
Django==4.2.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cors-headers==3.14.0
django-decouple==2.1
django-environ==0.10.0
django-graylog==0.7.0
django-rest-framework==0.1.0
django-restframework==0.0.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.1
lxml==4.9.3
msgpack==1.0.5
oscrypto==1.3.0
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg
psycopg-binary
psycopg2
pycparser==2.21
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
PyJWT==2.6.0
pyOpenSSL==23.2.0
pypdf==3.7.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz
qrcode==7.4.2
requests==2.28.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.7.1
tzdata
tzlocal
uritools==4.0.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
xhtml2pdf==0.2.9
zope.interface==6.0
